+++ src/game/default/g_item.h
@@ -25,7 +25,7 @@
 
 #ifdef __GAME_LOCAL_H__
 
-	extern const uint16_t g_num_items;
+	extern const size_t g_num_items;
 	
 	const g_item_t *G_ItemList(void);
